basic engine cpu profiler working with static build, but not with DLL

+++ Excessive-Engine/Common/src/ITimer.h
@@ -14,5 +14,5 @@ public:
 	virtual double getSecondsPassed() = 0;
 
 	// get micro seconds passed since start() or reset()
-	virtual size_t getMicroSecondsPassed() = 0;
+	virtual double getMicroSecondsPassed() = 0;
 };
